requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
trafilatura>=1.12
lxml_html_clean>=0.4
pyyaml>=6.0
